gce-xfstests: add GCE_REPORT_FAIL_EMAIL
When this config option is set, test runs with *hard* failures or test
infra errors send their reports to GCE_REPORT_FAIL_EMAIL instead of to
GCE_REPORT_EMAIL. If this option is not set, all emails go to
GCE_REPORT_EMAIL.

When we generate the summary report, if --check_failures is set,
we create a report.failed file if problems were found. When we go to
send emails, we check for this file and then route the email
correspondingly. We also send emails to GCE_REPORT_FAIL_EMAIL if the
ltm server detects issues. Support is included for normal gce-xfstests
runs, ltm runs, and kcs builds.

Tested passing and failing tests:
  with --fail-email flag and with config param GCE_REPORT_FAIL_EMAIL
  with and without LTM
  with REPORT_EMAIL="" and REPORT_FAIL_EMAIL=<email>
  with mutliple emails in REPORT_FAIL_EMAIL

Ensured the following get routed as failure:
  broken kcs build via ltm --commit <bad sha>
  errors resulting in ltm machine reboots
  timeouts via LTM monitor timeout
  timeouts via watchdog timeout with and without LTM

Signed-off-by: Leah Rumancik <leah.rumancik@gmail.com>